backend/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
supabase>=2.0.0  # Supabase Python client (auth, db, storage, realtime)
asyncpg>=0.29.0  # Async PostgreSQL driver
psycopg2-binary>=2.9.9  # PostgreSQL adapter (fallback/sync)

# Redis
redis>=5.0.1
hiredis>=2.2.3  # Faster Redis parsing

# Message Queue
celery>=5.3.4
kombu>=5.3.4  # Celery messaging library

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6  # For form data parsing

# WebSocket Support
websockets>=12.0

# HTTP Client (for service-to-service calls if needed)
httpx>=0.25.2

# Environment & Configuration
python-dotenv>=1.0.0

# Development Tools (optional but recommended)
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
ruff>=0.1.6
